api/requirements.txt
openai
google-cloud-storage>=1.16.1